refactor: modifica el servicio de notificaciones y el controlador para mejorar su estructura y legibilidad.
